# src/Domivice.Users.Web/wwwroot/openapi-original.json
{
    "openapi": "3.0.2",
    "info": {
        "contact": {
            "name": "Domivice Development Team",
            "url": "https://development.domivice.com"
        },
        "description": "The users API",
        "title": "Users API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
...
    "paths": {
...
        "/v1/users/{userId}/business-hours": {
            "post": {
                "parameters": [
                    {
                        "description": "The string guid representation of the user id",
                        "explode": false,
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BusinessHoursCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessHours"
                                }
                            }
                        },
                        "description": "Response when the business hours have been created successfully."
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "There was an error with the request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not authenticated"
                    }
                },
                "summary": "Create Business Hours",
                "tags": ["Business Hours"]
            }
        },
        "/v1/users/{userId}/business-hours/{businessHourId}": {
            "delete": {
                "parameters": [
                    {
                        "description": "The string guid representation of the user id",
                        "explode": false,
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }, {
                        "description": "The string guid representation of the business hour id",
                        "explode": false,
                        "in": "path",
                        "name": "businessHourId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Response when the business hours is deleted"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not authenticated"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not allowed to perform action"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "The resource was not found"
                    }
                },
                "summary": "Delete Business Hours",
                "tags": ["Business Hours"]
            },
            "get": {
                "parameters": [
                    {
                        "description": "The string guid representation of the user id",
                        "explode": false,
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }, {
                        "description": "The string guid representation of the business hour id",
                        "explode": false,
                        "in": "path",
                        "name": "businessHourId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessHours"
                                }
                            }
                        },
                        "description": "Response when an business hours object is returned"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not authenticated"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "The resource was not found"
                    }
                },
                "summary": "Get Business Hours",
                "tags": ["Business Hours"]
            },
            "patch": {
                "parameters": [
                    {
                        "description": "The string guid representation of the user id",
                        "explode": false,
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }, {
                        "description": "The string guid representation of the business hour id",
                        "explode": false,
                        "in": "path",
                        "name": "businessHourId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BusinessHoursUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessHours"
                                }
                            }
                        },
                        "description": "Response when the business hours have been successfully updated."
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "There was an error with the request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not authenticated"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not allowed to perform action"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "The resource was not found"
                    }
                },
                "summary": "Patch Business Hours",
                "tags": ["Business Hours"]
            }
        },
...
    },
    "components": {
...
        "schemas": {
            "Error": {
                "description": "",
                "properties": {
                    "detail": {
                        "description": "A human-readable explanation of the error",
                        "type": "string"
                    },
                    "instance": {
                        "description": "A URI that identifies the specific occurrence of the error",
                        "type": "string"
                    },
                    "status": {
                        "description": "The HTTP response code",
                        "type": "string"
                    },
                    "title": {
                        "description": "A brief, human-readable message about the error",
                        "type": "string"
                    },
                    "type": {
                        "description": "A URI identifier that categorizes the error",
                        "type": "string"
                    },
                    "errors": {
                        "description": "The validation errors",
                        "items": {
                            "$ref": "#/components/schemas/KeyValuePair"
                        },
                        "type": "array"
                    },
                    "extensions": {
                        "description": "The error extensions",
                        "items": {
                            "$ref": "#/components/schemas/KeyValuePair"
                        },
                        "type": "array"
                    }
                },
                "required": ["detail", "instance", "title", "type"],
                "type": "object"
            },
            "KeyValuePair": {
                "description": "Carries key values pair",
                "type": "object"
            },
...
            "BusinessHours": {
                "description": "Carries information about individual or agency openning hours",
                "example": {
                    "validThrough": "2000-01-23T04:56:07.000+00:00",
                    "timeSlots": [
                        {
                            "dayOfWeek": "monday",
                            "startTime": "2000-01-23T04:56:07.000+00:00",
                            "endTime": "2000-01-23T04:56:07.000+00:00"
                        }, {
                            "dayOfWeek": "monday",
                            "startTime": "2000-01-23T04:56:07.000+00:00",
                            "endTime": "2000-01-23T04:56:07.000+00:00"
                        }
                    ],
                    "validFrom": "2000-01-23T04:56:07.000+00:00",
                    "id": "id",
                    "userId": "userId"
                },
                "properties": {
                    "validFrom": {
                        "description": "The date when the item becomes valid.",
                        "format": "date-time",
                        "type": "string"
                    },
                    "validThrough": {
                        "description": "The date after when the item is not valid.",
                        "format": "date-time",
                        "type": "string"
                    },
                    "id": {
                        "description": "The opening hours record id",
                        "type": "string"
                    },
                    "userId": {
                        "description": "The user id",
                        "type": "string"
                    },
                    "timeSlots": {
                        "description": "The time slots of the business hours for a given week day",
                        "items": {
                            "$ref": "#/components/schemas/TimeSlot"
                        },
                        "type": "array"
                    }
                },
                "required": ["id", "timeSlots", "userId", "validFrom", "validThrough"],
                "type": "object"
            },
...
            "TimeSlot": {
                "description": "Carries information about open and closing time",
                "example": {
                    "dayOfWeek": "monday",
                    "startTime": "2000-01-23T04:56:07.000+00:00",
                    "endTime": "2000-01-23T04:56:07.000+00:00"
                },
                "properties": {
                    "endTime": {
                        "description": "The closing time",
                        "format": "date-time",
                        "type": "string"
                    },
                    "startTime": {
                        "description": "The opening time",
                        "format": "date-time",
                        "type": "string"
                    },
                    "dayOfWeek": {
                        "description": "The day of week number",
                        "enum": [
                            "monday", "tuesday", "wednesday", "thursday", "friday", "saturday", "sunday",
                            "publicHoliday"
                        ],
                        "type": "string"
                    }
                },
                "required": ["dayOfWeek", "endTime", "startTime"],
                "type": "object"
            },
            "BusinessHoursCreate": {
                "description": "Carries information about individual or agency openning hours",
                "example": {
                    "validThrough": "2000-01-23T04:56:07.000+00:00",
                    "timeSlots": [
                        {
                            "dayOfWeek": "monday",
                            "startTime": "2000-01-23T04:56:07.000+00:00",
                            "endTime": "2000-01-23T04:56:07.000+00:00"
                        }, {
                            "dayOfWeek": "monday",
                            "startTime": "2000-01-23T04:56:07.000+00:00",
                            "endTime": "2000-01-23T04:56:07.000+00:00"
                        }
                    ],
                    "validFrom": "2000-01-23T04:56:07.000+00:00"
                },
                "properties": {
                    "validFrom": {
                        "description": "The date when the item becomes valid.",
                        "format": "date-time",
                        "type": "string"
                    },
                    "validThrough": {
                        "description": "The date after when the item is not valid.",
                        "format": "date-time",
                        "type": "string"
                    },
                    "timeSlots": {
                        "description": "The time slots of the business hours for a given week day",
                        "items": {
                            "$ref": "#/components/schemas/TimeSlot"
                        },
                        "type": "array"
                    }
                },
                "required": ["timeSlots", "validFrom", "validThrough"],
                "type": "object"
            },
            "BusinessHoursUpdate": {
                "description": "Carries information about individual or agency openning hours",
                "example": {
                    "validThrough": "2000-01-23T04:56:07.000+00:00",
                    "timeSlots": [
                        {
                            "dayOfWeek": "monday",
                            "startTime": "2000-01-23T04:56:07.000+00:00",
                            "endTime": "2000-01-23T04:56:07.000+00:00"
                        }, {
                            "dayOfWeek": "monday",
                            "startTime": "2000-01-23T04:56:07.000+00:00",
                            "endTime": "2000-01-23T04:56:07.000+00:00"
                        }
                    ],
                    "validFrom": "2000-01-23T04:56:07.000+00:00"
                },
                "properties": {
                    "validFrom": {
                        "description": "The date when the item becomes valid.",
                        "format": "date-time",
                        "type": "string"
                    },
                    "validThrough": {
                        "description": "The date after when the item is not valid.",
                        "format": "date-time",
                        "type": "string"
                    },
                    "timeSlots": {
                        "description": "The time slots of the business hours for a given week day",
                        "items": {
                            "$ref": "#/components/schemas/TimeSlot"
                        },
                        "type": "array"
                    }
                },
                "type": "object"
            },
...
        }
    }
}